+++ R/zzz.R
@@ -4,4 +4,5 @@
   # If user unloads/detaches package make sure that these values are erased
   Sys.setenv("QUALTRICS_BASE_URL" = "")
   Sys.setenv("QUALTRICS_API_KEY" = "")
+  Sys.setenv("QUALTRICS_DIRECTORY_ID" = "")
 }